Move Callback struct to a separated module
