+++ src/lib.rs
@@ -63,6 +63,7 @@ pub mod html;
 pub mod prelude;
 pub mod services;
 pub mod virtual_dom;
+pub mod callback;
 
 /// Initializes yew framework. It should be called first.
 pub fn initialize() {
