+++ src/services/timeout.rs
@@ -2,7 +2,7 @@
 //! send a messages when timeout elapsed.
 
 use super::{to_ms, Task};
-use html::Callback;
+use callback::Callback;
 use std::time::Duration;
 use stdweb::Value;
